requirements.txt
Flask==2.3.0
qrcode==7.4.2
pillow==10.2.0
python-dotenv==1.1.1
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1